image-resin.bbclass: Add sanity check to prevent kernel in boot part
The kernel is now present in rootfs/boot/KERNEL instead of in the boot
partition.

Add a sanity check to prevent the kernel to be in boot partition files.

Cannot add the sanity check in resin-sanity.bbclass as those handlers
run at an early stage and RESIN_BOOT_PARTITION_FILES populates at a
later stage in multiple bbappends.

Signed-off-by: Zubair Lutfullah Kakakhel <zubair@resin.io>