resin-init-flasher: Give a chance for openvpn to connect
On boards where flashing takes a couple of seconds, the registration and
afterwards openvpn, don't have a change to complete. The effect of this
is that the dashboard will not show any feedback when the device is
connected to a backend. This change adds a timeout based loop in which
openvpn status is checked for 15 seconds. When that times out, the
flashing continues as normal assuming no connectivity or no backend
configured.

Fixes #1152

Change-type: patch
Changelog-entry: Fix dashboard feedback on fast flashing devices
Signed-off-by: Andrei Gherzan <andrei@resin.io>