resin-mounts-flasher: Move the rw copy of config.json out of /tmp
There are services which need to work on config.json but have a private
tmp mountpoint (see openvpn service). This makes the initial decision of
moving the config.json in /tmp not ideal. This commit defines:
* a new tmpfs mountpoint in /mnt/boottmp
* uses this new mountpoint for storing the rw copy of config.json

Fixes #1157

Change-type: patch
Changelog-entry: Move the rw copy of config.json out of /tmp for flasher
Signed-off-by: Andrei Gherzan <andrei@resin.io>