resin-device-uuid: Make this service part of basic target
Avoid racing with the os-config service by making this service part of
an earlier target: basic.target. The os-config service is part of
multi-user.target which runs after basic.target.

The problem at hand is that os-config reads and write the config.json
which races with the fact that uuid generation service writes as well to
the same config file. If this happens in between os-config's read and
write, we overwrite the changes done by the uuid service.

Fixes #1159

Change-type: patch
Changelog-entry: Avoid os-config / uuid service race
Signed-off-by: Andrei Gherzan <andrei@resin.io>