resin-device-register: Order this service before os-config
The problem at hand is that os-config reads and write the config.json
which races with the fact that registration service writes as well to
the same config file. If this happens in between os-config's read and
write, we overwrite the changes done by the registration service.

Change-type: patch
Changelog-entry: Avoid os-config / resin-device-register service race
Signed-off-by: Andrei Gherzan <andrei@resin.io>